requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
geographiclib==2.0
geopy==2.4.1
google==3.0.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
googleapis-common-protos==1.66.0
gps==3.19
gpsd-py3==0.3.0
grpcio==1.68.1
grpcio-status==1.68.1
h3==4.1.2
idna==3.10
numpy==2.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pytz==2024.2
requests==2.32.3
rsa==4.9
soupsieve==2.6
timezonefinder==6.5.7
urllib3==2.2.3
